requirements.txt
# Core scientific computing
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# Cheminformatics
rdkit-pypi>=2022.9.5

# 3D molecular visualization
py3Dmol>=2.0.0

# Web interface
streamlit>=1.25.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0

# Additional utilities
Pillow>=9.0.0
plotly>=5.10.0